requirements.txt
# Wind Turbine Platform — Python Dependencies
# Pin minor versions for reproducibility; allow patch updates.

# ── Kafka ──────────────────────────────────────────────────────────────────────
confluent-kafka==2.3.0

# ── Database ───────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ── Schema / Validation ────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1

# ── API ────────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Observability ──────────────────────────────────────────────────────────────
prometheus-client==0.20.0
structlog==24.1.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
httpx==0.27.0          # async test client for FastAPI

# ── Dev/Lint (not installed in Docker images — use locally) ───────────────────
# ruff==0.4.4
# mypy==1.10.0
